Add value_types trait in the rest tests, fix has_value_types

+++ test/parallel_api/algorithm/alg.modifying.operations/alg.reverse/reverse_copy.pass.cpp
@@ -103,7 +103,7 @@ main()
     test<std::int16_t, std::int8_t>();
     test<std::uint16_t, float32_t>();
     test<float64_t, std::int64_t>();
-    test<wrapper<float64_t>, wrapper<float64_t>>();
+    test<wrapper<float32_t>, wrapper<float32_t>>();
 
     return done();
 }
